refactor(ai): consolidate response_cache into cache.py
