+++ src/ai/clients/claude_client.py
@@ -1,29 +1,23 @@
 # src/ai/clients/claude_client.py
 # Claude (Anthropic) API client for generating JSON responses
 
-from functools import lru_cache
+from __future__ import annotations
 
 from .base import BaseClient
-from ..models import get_default_model
-from ..types import GenerateResult
 from ..utils import APICallContext
-from ...config.env_validator import validate_provider_env, get_missing_env_message
 from ...config.settings import settings_manager
-from ...core.exceptions import AIError, ConfigurationError
+from ...core.exceptions import AIError, ProviderError, RateLimitError
 
 
-# Anthropic Claude API client
+# * Anthropic Claude API client for JSON generation
 class ClaudeClient(BaseClient):
 
     provider_name = "anthropic"
+    required_env_vars = ["ANTHROPIC_API_KEY"]
 
-    # check if ANTHROPIC_API_KEY is available
-    def validate_credentials(self) -> None:
-        if not validate_provider_env("anthropic"):
-            raise ConfigurationError(get_missing_env_message("anthropic"))
-
-    # make Claude API call
+    # * Make Claude API call w/ JSON-only response mode
     def make_call(self, prompt: str, model: str) -> APICallContext:
+        import anthropic
         from anthropic import Anthropic
 
         client = Anthropic()
@@ -42,7 +36,34 @@ class ClaudeClient(BaseClient):
                 ],
             )
         except Exception as e:
-            raise AIError(f"Anthropic API error: {str(e)}")
+            # Safely check for provider-specific exception types (may not exist in mocks)
+            rate_limit_error = getattr(anthropic, "RateLimitError", None)
+            api_status_error = getattr(anthropic, "APIStatusError", None)
+            api_connection_error = getattr(anthropic, "APIConnectionError", None)
+
+            # Check for rate limit (429)
+            if rate_limit_error and isinstance(e, rate_limit_error):
+                raise RateLimitError(
+                    f"Anthropic rate limit exceeded: {e}",
+                    provider="anthropic",
+                    retry_after=getattr(e, "retry_after", None),
+                ) from e
+            # Check for API errors (4xx/5xx)
+            if api_status_error and isinstance(e, api_status_error):
+                status_code = getattr(e, "status_code", "unknown")
+                message = getattr(e, "message", str(e))
+                raise ProviderError(
+                    f"Anthropic API error ({status_code}): {message}",
+                    provider="anthropic",
+                ) from e
+            # Check for connection errors
+            if api_connection_error and isinstance(e, api_connection_error):
+                raise ProviderError(
+                    f"Anthropic connection error: {e}",
+                    provider="anthropic",
+                ) from e
+            # Fallback for unexpected errors
+            raise AIError(f"Anthropic API error: {e}") from e
 
         # extract text from response (process text blocks only & skip tool blocks)
         raw_text = ""
@@ -51,21 +72,3 @@ class ClaudeClient(BaseClient):
                 raw_text += content_block.text
 
         return APICallContext(raw_text=raw_text, provider_name="anthropic", model=model)
-
-
-# =============================================================================
-# Module-level API (backward compatibility + lazy singleton)
-# =============================================================================
-
-
-# get lazily-initialized singleton client
-@lru_cache(maxsize=1)
-def _get_client() -> ClaudeClient:
-    return ClaudeClient()
-
-
-# generate JSON response using Claude API
-# returns GenerateResult w/ success/failure status & data
-def run_generate(prompt: str, model: str | None = None) -> GenerateResult:
-    resolved_model = model or get_default_model("anthropic")
-    return _get_client().run_generate(prompt, resolved_model)
