+++ src/ai/clients/factory.py
@@ -1,39 +1,33 @@
 # src/ai/clients/factory.py
 # AI client factory for routing to appropriate provider based on model
 
+from __future__ import annotations
+
 from typing import Callable, Type
 
 from ..provider_validator import validate_model, get_model_error_message
-from ..models import resolve_model_alias
+from ..models import ModelRegistry
 from ..types import GenerateResult
 from .base import BaseClient
 
 
-# =============================================================================
-# Client Registry
-# =============================================================================
-
-# * Lazy client factories - return client class, instantiate on demand
-# Tests can monkeypatch these entries to inject mock clients
-
-
+# lazy client factory for OpenAI (tests can monkeypatch this)
 def _get_openai_client_class() -> Type[BaseClient]:
     from .openai_client import OpenAIClient
 
     return OpenAIClient
 
-
+# lazy client factory for Anthropic (tests can monkeypatch this)
 def _get_anthropic_client_class() -> Type[BaseClient]:
     from .claude_client import ClaudeClient
 
     return ClaudeClient
 
-
+# lazy client factory for Ollama (tests can monkeypatch this)
 def _get_ollama_client_class() -> Type[BaseClient]:
     from .ollama_client import OllamaClient
 
     return OllamaClient
-
 
 # * Registry mapping provider IDs to client factory functions
 CLIENT_REGISTRY: dict[str, Callable[[], Type[BaseClient]]] = {
@@ -42,14 +36,8 @@ CLIENT_REGISTRY: dict[str, Callable[[], Type[BaseClient]]] = {
     "ollama": _get_ollama_client_class,
 }
 
-
-# =============================================================================
-# Factory Function
-# =============================================================================
-
-
+# * Generate JSON response using appropriate AI client based on model
 def run_generate(prompt: str, model: str) -> GenerateResult:
-    # Generate JSON response using appropriate AI client based on model.
     # validate model & determine provider
     valid, provider = validate_model(model)
 
@@ -58,7 +46,7 @@ def run_generate(prompt: str, model: str) -> GenerateResult:
         return GenerateResult(success=False, error=error_msg)
 
     # resolve alias to full model name
-    resolved_model = resolve_model_alias(model)
+    resolved_model = ModelRegistry.resolve_alias(model)
 
     # get client factory from registry
     client_factory = CLIENT_REGISTRY.get(provider)  # type: ignore[arg-type]
