+++ src/ai/clients/ollama_client.py
@@ -1,8 +1,9 @@
 # src/ai/clients/ollama_client.py
 # Ollama API client for generating JSON responses using local models
 
+from __future__ import annotations
+
 from functools import lru_cache
-from typing import List
 
 from .base import BaseClient
 from ..cache import AICache
@@ -10,26 +11,26 @@ from ..models import get_default_model
 from ..types import GenerateResult, OllamaStatus
 from ..utils import APICallContext
 from ...config.settings import settings_manager
-from ...core.exceptions import AIError
-from ...core.debug import debug_ai, debug_error, debug_api_call
+from ...core.exceptions import AIError, ProviderError
+from ...core.output import get_output_manager
 
 
-# Ollama API client for local model inference
+# * Ollama API client for local model inference
 class OllamaClient(BaseClient):
 
     provider_name = "ollama"
 
-    # check Ollama server status before making API call
+    # * Check Ollama server status before making API call
     def preflight(self) -> None:
         status = self._check_ollama_status(with_debug=True)
         if not status.available:
             raise AIError(f"Ollama server error: {status.error}")
 
-    # Ollama doesn't require API key - handled by preflight
+    # Ollama doesn't require API key - validation handled by preflight
     def validate_credentials(self) -> None:
         pass
 
-    # validate model exists in Ollama's local model list
+    # * Validate model exists in Ollama's local model list
     def validate_model(self, model: str) -> str:
         status = self._check_ollama_status()
 
@@ -42,15 +43,17 @@ class OllamaClient(BaseClient):
 
         return model
 
-    # make Ollama API call
+    # * Make Ollama API call w/ JSON-only response mode
     def make_call(self, prompt: str, model: str) -> APICallContext:
         import ollama  # type: ignore
 
         settings = settings_manager.load()
+        output = get_output_manager()
 
-        debug_api_call("Ollama", model, len(prompt))
-        debug_ai(
-            f"Making Ollama API call with model: {model}, temperature: {settings.temperature}"
+        output.debug(f"Ollama API call - Model: {model}, Prompt: {len(prompt)} chars", "API")
+        output.debug(
+            f"Making Ollama API call with model: {model}, temperature: {settings.temperature}",
+            "AI",
         )
 
         try:
@@ -70,19 +73,34 @@ class OllamaClient(BaseClient):
             )
 
             raw_text = response.get("message", {}).get("content", "")
-            debug_ai(f"Received response from Ollama: {len(raw_text)} characters")
+            output.debug(f"Received response from Ollama: {len(raw_text)} characters", "AI")
 
             return APICallContext(
                 raw_text=raw_text, provider_name="ollama", model=model
             )
 
         except Exception as e:
-            debug_error(e, "Ollama API call")
-            raise AIError(
-                f"Ollama API error: {str(e)}. Model: {model}. Check if Ollama is running & model is properly installed."
+            output.debug(
+                f"Ollama API call - Exception: {type(e).__name__}: {str(e)}", "ERROR"
             )
+            # Check for ResponseError (Ollama's main error type)
+            if hasattr(ollama, "ResponseError") and isinstance(e, ollama.ResponseError):
+                raise ProviderError(
+                    f"Ollama API error: {e}",
+                    provider="ollama",
+                ) from e
+            # Check for connection errors
+            if isinstance(e, (ConnectionError, OSError)):
+                raise ProviderError(
+                    f"Ollama connection failed: {e}. Check if Ollama is running.",
+                    provider="ollama",
+                ) from e
+            # Fallback for unexpected errors
+            raise AIError(
+                f"Ollama API error: {e}. Model: {model}. Check if Ollama is running & model is installed."
+            ) from e
 
-    # check & cache Ollama server status
+    # check & cache Ollama server status (returns available models & error)
     def _check_ollama_status(self, *, with_debug: bool = False) -> OllamaStatus:
         # check cache first
         if AICache.is_ollama_cached():
@@ -97,15 +115,17 @@ class OllamaClient(BaseClient):
         try:
             import ollama  # type: ignore
 
+            output = get_output_manager()
             if with_debug:
-                debug_ai("Checking Ollama server availability...")
+                output.debug("Checking Ollama server availability...", "AI")
 
             response = ollama.list()
             models = [m.model for m in response.models if m.model]
 
             if with_debug:
-                debug_ai(
-                    f"Ollama server available - found {len(models)} models: {', '.join(models)}"
+                output.debug(
+                    f"Ollama server available - found {len(models)} models: {', '.join(models)}",
+                    "AI",
                 )
 
             AICache.set_ollama_status(models)
@@ -113,16 +133,15 @@ class OllamaClient(BaseClient):
 
         except Exception as e:
             if with_debug:
-                debug_error(e, "Ollama server check")
+                output = get_output_manager()
+                output.debug(
+                    f"Ollama server check - Exception: {type(e).__name__}: {str(e)}",
+                    "ERROR",
+                )
 
             error_msg = f"Ollama server connection failed: {str(e)}. Please ensure Ollama is running locally."
             AICache.set_ollama_status(None, error_msg)
             return OllamaStatus(available=False, models=[], error=error_msg)
-
-
-# =============================================================================
-# Module-level API (backward compatibility + lazy singleton)
-# =============================================================================
 
 
 # get lazily-initialized singleton client
@@ -130,47 +149,21 @@ class OllamaClient(BaseClient):
 def _get_client() -> OllamaClient:
     return OllamaClient()
 
-
-# generate JSON response using Ollama API
-# returns GenerateResult w/ success/failure status & data
+# * Generate JSON response using Ollama API
 def run_generate(prompt: str, model: str | None = None) -> GenerateResult:
     resolved_model = model or get_default_model("ollama")
     return _get_client().run_generate(prompt, resolved_model)
 
-
-# =============================================================================
-# Backward-compatibility wrappers (used by other modules)
-# =============================================================================
-
-
-# reset the Ollama cache
-def reset_cache() -> None:
-    AICache.invalidate_provider("ollama")
-    _get_client.cache_clear()
-
-
 # check Ollama server status (cached)
 def check_ollama_status(*, with_debug: bool = False) -> OllamaStatus:
     return _get_client()._check_ollama_status(with_debug=with_debug)
-
-
-# check if Ollama server is running & accessible
-def is_ollama_available() -> bool:
-    return check_ollama_status().available
-
 
 # check if Ollama server is running & return detailed error if not
 def check_ollama_with_error() -> tuple[bool, str]:
     status = check_ollama_status(with_debug=True)
     return status.available, status.error
 
-
-# get list of available local models from Ollama
-def get_available_models() -> List[str]:
-    return check_ollama_status().models
-
-
 # get list of available local models w/ detailed error reporting
-def get_available_models_with_error() -> tuple[List[str], str]:
+def get_available_models_with_error() -> tuple[list[str], str]:
     status = check_ollama_status(with_debug=True)
     return status.models, status.error
