+++ src/ai/models.py
@@ -1,19 +1,10 @@
 # src/ai/models.py
-# AI model catalog & aliases for Loom CLI (OpenAI, Anthropic & Ollama)
-#
-# This module provides:
-# - Model name catalogs (OPENAI_MODELS, CLAUDE_MODELS, etc.)
-# - Model aliases for user-friendly short names
-# - Default model configuration per provider
-# - Alias resolution
-#
-# Provider validation logic is in provider_validator.py
+# AI model catalog, aliases & registry for OpenAI, Anthropic & Ollama providers
 
-from typing import List, Optional
-
+from __future__ import annotations
 
 # * Supported OpenAI models
-OPENAI_MODELS: List[str] = [
+OPENAI_MODELS: list[str] = [
     "gpt-5",
     "gpt-5-mini",
     "gpt-5-nano",
@@ -22,7 +13,7 @@ OPENAI_MODELS: List[str] = [
 ]
 
 # * Supported Claude models (Anthropic)
-CLAUDE_MODELS: List[str] = [
+CLAUDE_MODELS: list[str] = [
     "claude-opus-4-1-20250805",
     "claude-opus-4-20250514",
     "claude-sonnet-4-20250514",
@@ -32,7 +23,7 @@ CLAUDE_MODELS: List[str] = [
 ]
 
 # * Combined supported models list (static models only, not Ollama)
-SUPPORTED_MODELS: List[str] = OPENAI_MODELS + CLAUDE_MODELS
+SUPPORTED_MODELS: list[str] = OPENAI_MODELS + CLAUDE_MODELS
 
 # * Model aliases for user-friendly short names
 MODEL_ALIASES: dict[str, str] = {
@@ -53,7 +44,7 @@ MODEL_ALIASES: dict[str, str] = {
 }
 
 # * Model categories for user guidance
-MODEL_CATEGORIES: dict[str, List[str]] = {
+MODEL_CATEGORIES: dict[str, list[str]] = {
     "latest": [
         "gpt-5",
         "gpt-5-mini",
@@ -103,30 +94,22 @@ MODEL_METADATA: dict[str, tuple[str, str]] = {
     "claude-3-haiku-20240307": ("anthropic", "Claude 3 Haiku (fastest, economical)"),
 }
 
-
 def get_model_description(model: str) -> str:
-    # Get human-readable description for a model.
     if model in MODEL_METADATA:
         return MODEL_METADATA[model][1]
     return model
 
-
 def resolve_model_alias(model: str) -> str:
-    # Resolve model alias to full model name.
     # first check if it's already a valid full model name
     if model in SUPPORTED_MODELS:
         return model
-
     # then check if it's an alias
     if model in MODEL_ALIASES:
         return MODEL_ALIASES[model]
-
     # return original if no alias found (may be Ollama model)
     return model
 
-
-def get_default_model(provider: Optional[str] = None) -> str:
-    # Get the default model for a provider.
+def get_default_model(provider: str | None = None) -> str:
     if provider is None:
         # global default is OpenAI
         return DEFAULT_MODELS_BY_PROVIDER["openai"]
@@ -134,11 +117,74 @@ def get_default_model(provider: Optional[str] = None) -> str:
         provider, DEFAULT_MODELS_BY_PROVIDER["openai"]
     )
 
-
-def get_provider_for_model(model: str) -> Optional[str]:
-    # Get the provider ID for a known static model. Note: This only checks static model lists. For dynamic validation including Ollama, use provider_validator.validate_model().
+def get_provider_for_model(model: str) -> str | None:
+    # ? Note: This only checks static model lists. For dynamic validation including Ollama, use provider_validator.validate_model()
     if model in OPENAI_MODELS:
         return "openai"
     if model in CLAUDE_MODELS:
         return "anthropic"
     return None
+
+# * Centralized model validation & resolution registry
+class ModelRegistry:
+    # resolve alias to full model name
+    @classmethod
+    def resolve_alias(cls, model: str) -> str:
+        if model in SUPPORTED_MODELS:
+            return model
+        return MODEL_ALIASES.get(model, model)
+
+    # get provider ID for known static model
+    @classmethod
+    def get_provider(cls, model: str) -> str | None:
+        if model in OPENAI_MODELS:
+            return "openai"
+        if model in CLAUDE_MODELS:
+            return "anthropic"
+        return None
+
+    # check if model is in static lists (OpenAI or Anthropic)
+    @classmethod
+    def is_static_model(cls, model: str) -> bool:
+        return model in SUPPORTED_MODELS
+
+    # get default model for provider
+    @classmethod
+    def get_default(cls, provider: str | None = None) -> str:
+        if provider is None:
+            return DEFAULT_MODELS_BY_PROVIDER["openai"]
+        return DEFAULT_MODELS_BY_PROVIDER.get(
+            provider, DEFAULT_MODELS_BY_PROVIDER["openai"]
+        )
+
+    # get human-readable description for model
+    @classmethod
+    def get_description(cls, model: str) -> str:
+        if model in MODEL_METADATA:
+            return MODEL_METADATA[model][1]
+        return model
+
+    # validate static model & return provider if valid
+    @classmethod
+    def validate_static(cls, model: str) -> tuple[bool, str | None]:
+        resolved = cls.resolve_alias(model)
+        provider = cls.get_provider(resolved)
+        if provider is not None:
+            return True, provider
+        return False, None
+
+    # list all models for provider (or all if provider is None)
+    @classmethod
+    def list_models(cls, provider: str | None = None) -> list[str]:
+        if provider == "openai":
+            return OPENAI_MODELS.copy()
+        if provider == "anthropic":
+            return CLAUDE_MODELS.copy()
+        if provider is None:
+            return SUPPORTED_MODELS.copy()
+        return []
+
+    # list all model aliases
+    @classmethod
+    def list_aliases(cls) -> dict[str, str]:
+        return MODEL_ALIASES.copy()
