+++ src/ai/utils.py
@@ -1,15 +1,10 @@
 # src/ai/utils.py
-# Shared utility functions for AI clients
-#
-# Response processing uses 3 layers:
-# 1. strip_markdown_code_blocks() - clean raw text
-# 2. parse_json() - parse to dict, return (data, error)
-# 3. validate_and_extract() - validate structure & normalize
+# Shared utility functions for AI response processing w/ JSON parsing & key normalization
 
 import json
 import re
 from dataclasses import dataclass
-from typing import Any, Callable, Optional
+from typing import Any, Optional
 
 from .types import GenerateResult
 
@@ -36,19 +31,14 @@ SECTION_KEY_ALIASES: dict[str, str] = {
 
 
 def normalize_op_keys(op: dict[str, Any]) -> dict[str, Any]:
-    # Expand short keys to full names for downstream compatibility.
     return {OP_KEY_ALIASES.get(k, k): v for k, v in op.items()}
 
-
 def normalize_edits_response(edits: dict[str, Any]) -> dict[str, Any]:
-    # Normalize all ops in an edits response (expand short keys).
     if "ops" in edits and isinstance(edits["ops"], list):
         edits["ops"] = [normalize_op_keys(op) for op in edits["ops"]]
     return edits
 
-
 def _normalize_subsection(sub: Any) -> dict[str, Any]:
-    # Normalize a subsection entry (array format or dict).
     if isinstance(sub, list) and len(sub) >= 3:
         # array format: [name, start_line, end_line, optional_meta]
         result: dict[str, Any] = {
@@ -71,7 +61,6 @@ def _normalize_subsection(sub: Any) -> dict[str, Any]:
 
 
 def normalize_section_keys(section: dict[str, Any]) -> dict[str, Any]:
-    # Expand short section keys to full names.
     normalized: dict[str, Any] = {}
     for k, v in section.items():
         key = SECTION_KEY_ALIASES.get(k, k) if isinstance(k, str) else str(k)
@@ -83,15 +72,13 @@ def normalize_section_keys(section: dict[str, Any]) -> dict[str, Any]:
         ]
     return normalized
 
-
 def normalize_sections_response(data: dict[str, Any]) -> dict[str, Any]:
-    # Normalize all sections in a sectionizer response (expand short keys).
     if "sections" in data and isinstance(data["sections"], list):
         data["sections"] = [normalize_section_keys(s) for s in data["sections"]]
     return data
 
 
-# * Context object for API call results (used by process_json_response)
+# * Context object for API call results (used by BaseClient._process_response)
 @dataclass(slots=True)
 class APICallContext:
     raw_text: str  # raw response text from provider
@@ -99,13 +86,8 @@ class APICallContext:
     model: str  # model used for the call
 
 
-# =============================================================================
-# Layer 1: Text stripping
-# =============================================================================
-
-
+# strip markdown code blocks & thinking tokens from AI responses
 def strip_markdown_code_blocks(text: str) -> str:
-    # Strip markdown code blocks & thinking tokens from AI responses.
     # first strip thinking tokens if present
     thinking_pattern = r"<think>.*?</think>\s*(.*)"
     thinking_match = re.match(thinking_pattern, text.strip(), re.DOTALL)
@@ -122,13 +104,8 @@ def strip_markdown_code_blocks(text: str) -> str:
         return text.strip()
 
 
-# =============================================================================
-# Layer 2: JSON parsing
-# =============================================================================
-
-
+# * Parse JSON from text after stripping markdown (returns data, json_text, error_message)
 def parse_json(text: str) -> tuple[Optional[dict[str, Any]], str, str]:
-    # Parse JSON from text after stripping markdown. (data, json_text, error_message) - data: parsed dict on success, None on failure - json_text: the stripped text that was parsed - error_message: empty on success, error details on failure
     json_text = strip_markdown_code_blocks(text)
 
     try:
@@ -140,11 +117,7 @@ def parse_json(text: str) -> tuple[Optional[dict[str, Any]], str, str]:
         return None, json_text, error_msg
 
 
-# =============================================================================
-# Layer 3: Validation & extraction
-# =============================================================================
-
-
+# * Validate AI response structure & extract normalized data (raises JSONParsingError or AIError)
 def validate_and_extract(
     data: Any,
     raw_text: str,
@@ -158,7 +131,6 @@ def validate_and_extract(
     log_version_debug: bool = False,
     log_structure: bool = False,
 ) -> dict[str, Any]:
-    # Validate AI response structure & extract normalized data. Raises: JSONParsingError: if JSON parsing failed but we got a response AIError: if API call failed or response structure is invalid
     # ! import here to avoid circular dependency w/ core module
     from ..core.exceptions import JSONParsingError, AIError
     from ..core.debug import debug_ai
@@ -232,28 +204,7 @@ def validate_and_extract(
     return normalize_edits_response(data)
 
 
-# =============================================================================
-# High-level helpers
-# =============================================================================
-
-
-def process_json_response(
-    api_call: Callable[[str, str], APICallContext], prompt: str, model: str
-) -> GenerateResult:
-    # Process API response into GenerateResult w/ consistent JSON parsing. Used by clients to convert raw API call to GenerateResult.
-    ctx = api_call(prompt, model)
-    data, json_text, error = parse_json(ctx.raw_text)
-
-    if data is not None:
-        return GenerateResult(
-            success=True, data=data, raw_text=ctx.raw_text, json_text=json_text
-        )
-    else:
-        return GenerateResult(
-            success=False, raw_text=ctx.raw_text, json_text=json_text, error=error
-        )
-
-
+# * High-level helper: GenerateResult -> validated dict w/ structure validation
 def process_ai_response(
     result: GenerateResult,
     model: str,
@@ -264,7 +215,6 @@ def process_ai_response(
     log_version_debug: bool = False,
     log_structure: bool = False,
 ) -> dict[str, Any]:
-    # High-level helper: GenerateResult -> validated dict. Orchestrates the 3-layer processing: 1. (already done) strip & parse in GenerateResult 2. validate_and_extract for structure validation Raises: JSONParsingError: if JSON parsing failed AIError: if validation fails
     # if result already failed, extract error info
     if not result.success:
         # parse_json was already called, pass through the error
