+++ tests/unit/ai/clients/test_claude_client.py
@@ -8,7 +8,7 @@ import pytest
 from unittest.mock import patch, Mock
 import os
 
-# provide minimal anthropic module so client import succeeds even if not installed
+# Provide minimal anthropic module so client import succeeds even if not installed
 if "anthropic" not in sys.modules:
     dummy = types.ModuleType("anthropic")
 
@@ -18,17 +18,14 @@ if "anthropic" not in sys.modules:
     dummy.Anthropic = _Placeholder  # type: ignore
     sys.modules["anthropic"] = dummy
 
-from src.ai.clients.claude_client import run_generate, ClaudeClient, _get_client
+from src.ai.clients.claude_client import ClaudeClient
+from src.ai.types import GenerateResult
 from src.core.exceptions import AIError, ConfigurationError
 
-
-@pytest.fixture(autouse=True)
-def reset_client_cache():
-    # Reset the singleton client cache before each test.
-    _get_client.cache_clear()
-    yield
-    _get_client.cache_clear()
-
+# * Helper function to run generate using client directly (no more module-level API)
+def _run_generate(prompt: str, model: str) -> GenerateResult:
+    client = ClaudeClient()
+    return client.run_generate(prompt, model)
 
 # * Fake Anthropic response objects for testing
 class _FakeContentBlock:
@@ -36,11 +33,9 @@ class _FakeContentBlock:
         self.type = "text"
         self.text = text
 
-
 class _FakeResponse:
     def __init__(self, text: str):
         self.content = [_FakeContentBlock(text)]
-
 
 class _FakeMessages:
     def __init__(
@@ -54,13 +49,11 @@ class _FakeMessages:
             raise self._error
         return _FakeResponse(self._response_text or "{}")
 
-
 class _FakeAnthropic:
     def __init__(
         self, response_text: str | None = None, error: Exception | None = None
     ):
         self.messages = _FakeMessages(response_text, error)
-
 
 # * Test successful result normalization & JSON parsing
 @patch("anthropic.Anthropic")
@@ -71,12 +64,11 @@ def test_claude_success_normalized_result(mock_anthropic_class):
 
     mock_anthropic_class.return_value = _FakeAnthropic(json.dumps(payload))
 
-    result = run_generate("Parse resume", model="claude-sonnet-4-20250514")
+    result = _run_generate("Parse resume", "claude-sonnet-4-20250514")
     assert result.success is True
     assert result.data == payload
     assert result.raw_text
     assert result.json_text
-
 
 # * Test API error raised as AIError (caught by base class, returns error result)
 @patch("anthropic.Anthropic")
@@ -85,20 +77,18 @@ def test_claude_success_normalized_result(mock_anthropic_class):
 def test_claude_api_error_returns_error_result(mock_anthropic_class):
     mock_anthropic_class.return_value = _FakeAnthropic(None, RuntimeError("boom"))
 
-    result = run_generate("Parse resume", model="claude-sonnet-4-20250514")
+    result = _run_generate("Parse resume", "claude-sonnet-4-20250514")
     assert result.success is False
     assert "Anthropic API error" in result.error
-
 
 # * Test missing API key returns error result
 @patch.dict(os.environ, {}, clear=True)
 # * Verify claude missing api key returns error result
 def test_claude_missing_api_key_returns_error_result():
-    result = run_generate("Test prompt", model="claude-sonnet-4-20250514")
+    result = _run_generate("Test prompt", "claude-sonnet-4-20250514")
 
     assert result.success is False
     assert "ANTHROPIC_API_KEY" in result.error or "anthropic" in result.error.lower()
-
 
 # * Test markdown code block stripping
 @patch("anthropic.Anthropic")
@@ -110,13 +100,12 @@ def test_claude_strips_markdown_code_blocks(mock_anthropic_class):
 
     mock_anthropic_class.return_value = _FakeAnthropic(markdown_response)
 
-    result = run_generate("Test prompt", model="claude-sonnet-4-20250514")
+    result = _run_generate("Test prompt", "claude-sonnet-4-20250514")
 
     assert result.success is True
     assert result.raw_text == markdown_response
     assert result.json_text == json.dumps(payload)
     assert result.data == payload
-
 
 # * Test ClaudeClient class directly
 class TestClaudeClientClass:
